requirements.txt
pyyaml>=5.1
uproot>=3.0
matplotlib>=3.0
pandas>=0.24.0
scikit-learn>=0.20
tables>=3.4
numexpr>=2.6.5
h5py>=2.8.0
lz4>=2.1.5